Correct the behavior and return value of snprintf()
According to C99 7.19.6.5, it explicitly defines the behavior and
return value of snprintf(). However, the implementation should
consider the following description:

 - If n is zero, nothing is written.
 - Writes at most n bytes, including the null character.
 - On success, the return value should be the length of the
   entire converted string even if n is insufficient to store it.

While validating snprintf() in the built-in C library, it was found
that the behavior and/or the return value did not match the above
description.

Therefore, these changes fix the implementation of snprintf() to
comply with the specification, and the related test cases are also
adjusted to verify the behavior and return value of snprintf().