Refine the source_t structure and its functions for general use
Originally, source_t was defined as a structure to store the text of
source files, implemented as a dynamic array. However, these changes
rename this structure and its related functions to make them more
generic. They are no longer limited to storing source code but
suitable for any variable-length character data.

The structure is now renamed to strbuf_t, and can be used to create
growable buffers of characters.